requirements.txt
alembic==1.11.1
APScheduler==3.10.3
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
flask-sqlacodegen==2.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
inflect==6.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.25.0
openpyxl==3.1.2
pandas==2.0.3
protobuf==3.20.3
pycryptodome==3.18.0
pydantic==1.10.9
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.14
tabulate==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1

